polyhedral_separation_set — Function

polyhedral_separation_set(G::Graph{Directed}, C::Matrix{RealExpr}, i::Int64, j::Int64, L=Int64[])

Computes a description of the polyhedral set of all matrices C which satisfy the C*-separation [i _||_ j | L] with respect to G.

The graph G must be transitively closed.

polyhedral_generic_set — Method

polyhedral_generic_set(G::Graph{Directed}, C::Matrix{RealExpr})

Computes a description of the set of generic weight matrices for G. For these matrices no two distinct paths between any pair of nodes has the same weight.

maxoid_implication — Method

maxoid_implication(G::Graph{Directed}, P::Vector{CIStmt}, Q::Vector{CIStmt}; generic_only=false)

Test if every maxoid associated to G satisfies the implication and(P) => or(Q) (the local CI implication problem). If generic_only is true, then only generic maxoids for G are tested.

Returns a tuple consisting of a boolean to indicate if the implication is true, and if it is false, also a counterexample matrix.

The graph G must be transitively closed.

maxoid_implication — Method

maxoid_implication(n::Int, P::Vector{CIStmt}, Q::Vector{CIStmt}; generic_only=false)

Test if every maxoid on n nodes satisfies the implication and(P) => or(Q) (the global CI implication problem). If generic_only is true, then only those maxoids are tested which are generic for some graph.

Returns a tuple consisting of a boolean to indicate if the implication is true, and if it is false, also a graph and a corresponding counterexample matrix.

The graph G must be transitively closed.